pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "panel-twin"
version = "0.1.0"
description = "Digital panel twin simulator for pilot-to-scale economics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "pydantic",
  "matplotlib",
  "PyYAML",
  "streamlit",
]

[project.scripts]
twin-econ = "twin_econ.cli:main"

[tool.setuptools]
packages = ["twin_econ"]
